==> Jupyter_R_2023/2_R_global_coo.ipynb <==
library(dplyr)
library(GenomicFeatures)

# this scripts transform local (transcriptomic) coordinates of extensions into global (genomic) coordinates using .gtf file. 
txdb <- makeTxDbFromGFF('/home/alla/make_PhD_great_again/NTE/VERSION_since_May2021/data/gencode.v25.annotation.gtf')


# Known sole nonAUG: 5'UTR+nonAUG+dAUG1+4nt for the context 

cr <- read.csv('tmp_res/known_sole_nonAUG_local_coo.txt',  sep='\t', stringsAsFactors = FALSE)

cr

tr_names <- cr$tr_id
start <- 1 #it was zero  # cds_start_pos	utr5_start
end <- cr$dAUG1_start+7

exons <- exonsBy(txdb, by="tx", use.names=TRUE)[tr_names]
x <- GRanges(tr_names, IRanges(start=start, end=end)) #,  
out <- data.frame(pmapFromTranscripts(x, exons))
write.table(out, 'tmp_res/known_sole_nonAUG_global_coo.txt', sep='\t')

# All transcript with cdsATG 

cr <- read.csv('tmp_res/utr5_cdsATG_4nt_local.txt',  sep='\t', stringsAsFactors = FALSE)

head(cr)

tr_names <- cr$tr_id
start <- 1 #it was zero  # cds_start_pos	utr5_start
end <- cr$cds_start_pos+7

exons <- exonsBy(txdb, by="tx", use.names=TRUE)[tr_names]
x <- GRanges(tr_names, IRanges(start=start, end=end)) #,  
out <- data.frame(pmapFromTranscripts(x, exons))
write.table(out, 'tmp_res/utr5_cdsATG_4nt_global.txt', sep='\t')
