==> Jupyter_R_2023/6_secondary_structure_G_quandruplexes_Ala_TIS.ipynb <==
%matplotlib inline

import numpy as np
import collections
from collections import OrderedDict, Counter, defaultdict
import pandas as pd

import Bio
from Bio import SeqIO
from Bio.Seq import Seq

import seaborn as sns
import matplotlib.pyplot as plt
from bokeh.plotting import figure, output_file, show 
import matplotlib.patches as patches    
from matplotlib import colors


import glob

import subprocess
from subprocess import call

import re

import pickle

from Bio.SubsMat import MatrixInfo
import Bio.Data.CodonTable

from scipy import stats


from Bio.SubsMat import MatrixInfo
import Bio.Data.CodonTable

import joypy

import pyBigWig

path = '/home/alla/make_PhD_great_again/NTE/VERSION_since_May2021/'

# gencode 25
metadata_pc_g25 = pd.read_csv(path+'tmp_res/metadata_pc_g25.txt', sep='\t')
metadata_pc_g25['tr_id1'] = [x.split('.')[0] for x in metadata_pc_g25['tr_id'].tolist()] 

# gencode 41
metadata_pc_g41 = pd.read_csv('tmp_res/metadata_pc_g41.txt', sep='\t')
metadata_pc_g41['tr_id1'] = [x.split('.')[0] for x in metadata_pc_g41['tr_id'].tolist()] 
metadata_pc_g41[0:2]


# open metadata with scores and metrics
meta = pd.read_csv('/home/alla/make_PhD_great_again/NTE/VERSION_since_May2021/tmp_res/METADATA_CURRENT.txt', sep='\t')
meta[0:2]

# nonAUG TIS 

TIS_efficiency_df = pd.read_csv('/home/alla/make_PhD_great_again/NTE/VERSION_since_May2021/data/TIS_nonAUG_efficiency.txt', sep='\t')

TIS_efficiency_df['scaled_efficiency'] = TIS_efficiency_df['TIS Efficiency'] / TIS_efficiency_df['TIS Efficiency'].max()

TIS_efficiency_dict = dict(zip(TIS_efficiency_df['TIS Sequence'], TIS_efficiency_df['TIS Efficiency']))

TIS_efficiency_dict_scaled = dict(zip(TIS_efficiency_df['TIS Sequence'], TIS_efficiency_df['scaled_efficiency']))

TIS_efficiency_df.sort_values(by='TIS Efficiency', ascending=False).head(4)

# AUG TIS 

TIS_AUG_eff = pd.read_csv('/home/alla/make_PhD_great_again/NTE/VERSION_since_May2021/data/TIS_AUG_efficiency.txt', sep='\t')

TIS_AUG_eff['scaled_efficiency'] = TIS_AUG_eff['efficiency'] / TIS_AUG_eff['efficiency'].max()

print (TIS_AUG_eff.efficiency.max(), TIS_AUG_eff.efficiency.min())


TIS_AUG_eff_dict = dict(zip(TIS_AUG_eff['sequence'], TIS_AUG_eff['efficiency']))


TIS_AUG_eff_dict_scaled = dict(zip(TIS_AUG_eff['sequence'], TIS_AUG_eff['scaled_efficiency']))
                           
TIS_AUG_eff[0:5]

li = []

gene = 'CCDC8'

tr_id = 'ENST00000307522'

pre_st = '565_CTG'

li.append([gene, tr_id, int(pre_st.split('_')[0]), pre_st.split('_')[1]])

gene = 'EIF4G2'

tr_id = 'ENST00000339995'

pre_st = '493_GTG'

li.append([gene, tr_id, int(pre_st.split('_')[0]), pre_st.split('_')[1]])


gene = 'BAG1'

tr_id = 'ENST00000472232'

pre_st = '88_CTG'

li.append([gene, tr_id, int(pre_st.split('_')[0]), pre_st.split('_')[1]])


gene = 'GTF3A'

tr_id = 'ENST00000381140'

pre_st = '195_CTG'

li.append([gene, tr_id, int(pre_st.split('_')[0]), pre_st.split('_')[1]])


gene = 'DDX17'

tr_id = 'ENST00000396821'

pre_st = '101_CTG'

li.append([gene, tr_id, int(pre_st.split('_')[0]), pre_st.split('_')[1]])


gene = 'RNF187'

tr_id = 'ENST00000305943'

pre_st = '429_CTG'

li.append([gene, tr_id, int(pre_st.split('_')[0]), pre_st.split('_')[1]])



gene = 'NELFB'

tr_id = 'ENST00000343053'

pre_st = '194_CTG'

li.append([gene, tr_id, int(pre_st.split('_')[0]), pre_st.split('_')[1]])



gene = 'NPTXR'

tr_id = 'ENST00000333039'

pre_st = '125_CTG'

li.append([gene, tr_id, int(pre_st.split('_')[0]), pre_st.split('_')[1]])



gene = 'NPW'

tr_id = 'ENST00000329610'

pre_st = '383_CTG'

li.append([gene, tr_id, int(pre_st.split('_')[0]), pre_st.split('_')[1]])



gene = 'PIGX'

tr_id = 'ENST00000392391'

pre_st = '176_CTG'

li.append([gene, tr_id, int(pre_st.split('_')[0]), pre_st.split('_')[1]])




gene = 'PRPS1L1'

tr_id = 'ENST00000506618'

pre_st = '82_ACG'

li.append([gene, tr_id, int(pre_st.split('_')[0]), pre_st.split('_')[1]])



#gene = 'PRR16'

#tr_id = 'ENST00000446965'

#pre_st = '82_GTG'

#li.append([gene, tr_id, int(pre_st.split('_')[0]), pre_st.split('_')[1]])




gene = 'SLC22A17'

tr_id = 'ENST00000397260'

pre_st = '120_CTG'

li.append([gene, tr_id, int(pre_st.split('_')[0]), pre_st.split('_')[1]])



gene = 'SP3'

tr_id = 'ENST00000418194'

pre_st = '385_ATA'

li.append([gene, tr_id, int(pre_st.split('_')[0]), pre_st.split('_')[1]])




gene = 'SWI5'

tr_id = 'ENST00000608796'

pre_st = '120_CTG'

li.append([gene, tr_id, int(pre_st.split('_')[0]), pre_st.split('_')[1]])



gene = 'TXNRD3'

tr_id = 'ENST00000524230'

pre_st = '166_CTG'

li.append([gene, tr_id, int(pre_st.split('_')[0]), pre_st.split('_')[1]])


gene = 'R3HCC1'

tr_id = 'ENST00000265806'

pre_st = '269_CTG'

li.append([gene, tr_id, int(pre_st.split('_')[0]), pre_st.split('_')[1]])

df = pd.DataFrame(li, columns = ['gene', 'tr_id1', 'start', 'codon'])

metadata_pc_g25['tr_id1'] = [x.split('.')[0] for x in metadata_pc_g25['tr_id'].tolist()]

df[0:2]

li = []
for el in df.to_numpy():
    seq = metadata_pc_g25[metadata_pc_g25['tr_id1'] == el[1]].iloc[0].transcript_seq
    start = el[2]
    nonAUG_context = seq[start-5:start+6]
    li.append([el[0], el[1], el[2], el[3], nonAUG_context])
    
    # nonAUG TIS score 

pd.DataFrame(li)

df

# TIS score of sink ORFs and nonAUG starts 

could be actually like a two-sided barplot with showing top-score as dashed line for both types of codons 

# Any polyAla downstream of nonAUG starts? 

# Any quadruplexes?
